+++ yudao-module-mini/src/main/java/cn/iocoder/yudao/module/custom/controller/admin/contract/vo/ContractRespVO.java
@@ -2,6 +2,8 @@ package cn.iocoder.yudao.module.custom.controller.admin.contract.vo;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.util.*;
 import org.springframework.format.annotation.DateTimeFormat;
 import java.time.LocalDateTime;
@@ -71,11 +73,11 @@ public class ContractRespVO {
 
     @Schema(description = "金额")
     @ExcelProperty("金额")
-    private Long salary;
+    private BigDecimal salary;
 
     @Schema(description = "费率")
     @ExcelProperty("费率")
-    private Long tariff;
+    private BigDecimal tariff;
 
     @Schema(description = "借条附件")
     @ExcelProperty("借条附件")
@@ -83,10 +85,10 @@ public class ContractRespVO {
 
     @Schema(description = "利息")
     @ExcelProperty("利息")
-    private Long interest;
+    private BigDecimal interest;
 
     @Schema(description = "已还金额")
     @ExcelProperty("已还金额")
-    private Long refund;
+    private BigDecimal refund;
 
 }
